sawyer_middleware/rest-server/requirements.txt
flask
flask-socketio
python2-secrets
qi==3.0.0
gunicorn==20.1.0
eventlet==0.30.2
numpy
matplotlib
Pillow
